@@ -2,9 +2,9 @@
 name = "django-minio-storage"
 description = "Django file storage using the minio python client"
 license = {file = "LICENSE"}
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
-  "django>=3.2",
+  "django>=4.2",
   "minio>=7.1.16",
 ]
 classifiers=[
@@ -41,7 +41,7 @@ write_to_template = '__version__ = "{version}"'
 tag_regex =  "^v(?P<prefix>v)?(?P<version>[^\\+]+)(?P<suffix>.*)?$"
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 88
 
 [tool.ruff.lint]
